rag-agent-llama/requirements.txt
psycopg==3.1.19
colorama==0.4.6
chromadb==0.5.0
tqdm==4.66.4
ollama==0.2.1
python-dotenv==1.0.0




